+++ src/roq/binance/order_entry_ws.cpp
@@ -1187,7 +1187,7 @@ void OrderEntryWS::update_helper(
 }
 
 template <typename... Args>
-void OrderEntryWS::operator()(Trace<oms::Response> const &event, uint8_t user_id, uint32_t order_id, Args &&...args) {
+void OrderEntryWS::operator()(Trace<oms::Response> const &event, uint8_t user_id, uint64_t order_id, Args &&...args) {
   auto &[trace_info, response] = event;
   if (shared_.update_order(
           user_id,
